Add tests for uv_guess_handle.
This adds relatively comprehensive tests for UNIX platforms,
however the Windows testing is currently light-on. Someone
with strong windows experience may be able to improve this.

The current #ifdef conditional compilation approach is probably
not ideal. It may be simpler to have a seperate test function
for windows and UNIX than the current approach.